requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
face-recognition>=1.3.0
basicsr>=1.4.0
gradio>=3.0.0
